Material Calendar Library
Removed all unused imports ,  commented Code segments and changed Button to Floating Action Button.

+++ HarmonyOS/Libraries/MaterialCalendar/materialcalendar/src/main/java/com/jmavarez/materialcalendar/Util/CanvasHelper.java
@@ -5,7 +5,7 @@ import ohos.app.Context;
 
 public class CanvasHelper {
 
-    public static float dpToPx(Context context, int dp) {
+    public static int dpToPx(Context context, int dp) {
         return dp * context.getResourceManager().getDeviceCapability().screenDensity / 160;
     }
